# [RFC] Opt/hash op