--- a/paddle/fluid/operators/hash_op.h
+++ b/paddle/fluid/operators/hash_op.h
@@ -15,6 +15,8 @@ limitations under the License. */
 #pragma once
 
 extern "C" {
+#define XXH_INLINE_ALL
+#define XXH_FORCE_MEMORY_ACCESS 1
 #include <xxhash.h>
 }
 #include <libdivide.h>